fix: start Storybook from root of monorepo for mdx support
Storybook does not currently support loading `.mdx` files from outside the package
or from inside node_modules/. This workaround is to start Storybook from the root
package of the monorepo that contains all `.mdx` files, so there is no issue.

Ideally we find a way to tweak the babel config to enable compilation of `.mdx`
files from any directory we fancy.