ci: tweak ESLint rules for React component library
Didn't want to enable Jest globals for all files, so I only enable it for this package.
Ideally the Jest globals would only be available to *.test.{js,jsx,ts,tsx},
but that seems like too much hassle to maintain.

Note: it is strange that `"react/react-in-jsx-scope": "off"` here is necessary, since
it should already work from the setting in /eslint.react.json