docs: temporarily disable Calendar stories
The @utrecht/component-library-react build does not include the calendar component in the dist/. A while back `export Calendar` had been removed because of the peer dependency on date-fns, but the rollup config does not have a separate build target for the standalone Calendar component. Remove the Story for Calendar until the build has been fixed.

Necessary for migration from Webpack to Vite