fix: correct the radio group markup
Putting the radio input element inside of the label breaks the css grid
grid-area because the parent has 'display: inline' (by default, because
it's a label).

Moving the input before the label paragraph wrapper fixes this, putting
the input and label both in the correct grid slot, which causes text
to wrap (in the label) to stay within its own grid area.

Label click still focuses the input element, on the condition that the
htmlFor and id attributes are properly linked.

This patch also adds a long, wrapping label for the label to get
visual regression testing on this aspect.