build(storybook-react): add missing dependencies
There is an issue with CSS dependencies. If a CSS package depends on another package, you must install both packages when using the main package.
For instance, if you need to use the button-css in Storybook, you must also install the `@utrecht/focus-ring-css` package because the button depends on it.