fix: component-library-react issues
Move `packages/component-library-react/packages/<component>` to
`packages/components-react/<component>`

Make these components individually build and pass their own tests.

Fix imports of `@utrecht/button-react` and remove the `src/index` part
that would make them part of `"rootDir"`.

Introduce `tsconfig.build.json` that extends the normal `tsconfig.json`.
The main difference is that this sets the `compilerOption` `composite`
to `true` which is needed to use the project as a reference in another
project. This was done in `@utrecht/button-react`,
`@utrecht/calendar-react` and `@utrecht/component-library-react` that
are all referenced from elsewhere in the codebase.