docs: apply alphabetical sorting to storybooks with components
Feedback from community in designer open hour, that components are already hard to find in Storybook when filtering, so users will search for components alphabetically.

In some cases, components were not ordered alphabetically (ie Acccordion being at the near bottom of the CSS Components Storybook). We needed to set sorting explicitly to alphabetical in addition to the custom storySort order.